feat: otp 사용자 쿠폰 적립 소진 요청
ISSUE: #34

+++ src/main/java/kr/kro/deom/domain/otp/entity/OtpUsage.java
@@ -30,7 +30,7 @@ public class OtpUsage {
     private OtpType type; // stamp | deom
 
     @Column(name = "deom_id")
-    private Integer deomId;
+    private Long deomId;
 
     @Column(name = "used_stamp_amount")
     private Integer usedStampAmount;
